+++ Myslot/Myslot.lua
@@ -398,7 +398,7 @@ function MySlot:Export(opt)
     end
 
     msg.petslot = {}
-    if not opt.ignorePetActionBar then
+    if not opt.ignorePetActionBar and IsPetActive() then
         for i = 1, NUM_PET_ACTION_SLOTS, 1 do
             local m = self:GetPetActionInfo(i)
             if m then
@@ -737,6 +737,14 @@ function MySlot:RecoverData(msg, opt)
                     if slotType == MYSLOT_SPELL then
                         PickupSpell(index)
 
+                        -- fallback option - try to get base spell
+                        if not GetCursorInfo() and FindBaseSpellByID then
+                            local baseSpellId = FindBaseSpellByID(index)
+                            if baseSpellId then
+                                PickupSpell(baseSpellId)
+                            end
+                        end
+
                         -- try if override
                         if not GetCursorInfo() then
                             if spellOverride[index] then
@@ -751,6 +759,7 @@ function MySlot:RecoverData(msg, opt)
                                 PickupSpell(spellName)
                             end
                         end
+
 
                         if not GetCursorInfo() then
                             MySlot:Print(L["Ignore unlearned skill [id=%s], %s"]:format(index, GetSpellLink(index) or ""))
@@ -848,7 +857,7 @@ function MySlot:RecoverData(msg, opt)
     end
 
 
-    if not opt.actionOpt.ignorePetActionBar then
+    if not opt.actionOpt.ignorePetActionBar and IsPetActive() then
         local pettoken = {}
         for i = 1, NUM_PET_ACTION_SLOTS, 1 do
             local name, _, isToken = GetPetActionInfo(i)