+++ Myslot/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Myslot
 
-## [v5.22.2](https://github.com/tg123/myslot/tree/v5.22.2) (2024-11-02)
-[Full Changelog](https://github.com/tg123/myslot/commits/v5.22.2) 
+## [v5.22.4](https://github.com/tg123/myslot/tree/v5.22.4) (2025-01-17)
+[Full Changelog](https://github.com/tg123/myslot/compare/v5.22.3...v5.22.4) 
 
-- bump interface versions for compatibility updates  
+- #76: Add FindBaseSpell fallback option (#77)  
+    * #76: Add FindBaseSpell fallback option  
+    * Update Myslot.lua  
+    Co-authored-by: Boshi Lian <farmer1992@gmail.com>  
+    ---------  
+    Co-authored-by: hewozuoai <sapc@tuta.io>  
+    Co-authored-by: Boshi Lian <farmer1992@gmail.com>  
+- fix: ensure pet action bar is only ignored when no active pets